moved widgets to widgets folder & contexts to context folder
