+++ src/components/CSIController/widgets/InputEnum.tsx
