+++ src/components/CSIController/widgets/Rectangle.tsx
